+++ src/api/display/layout.rs
@@ -9,7 +9,7 @@ use super::{
 pub struct Layout {
     /// A Map is a 2D vector, where the Vec<_> are rows and Vec<Vec<_>> are items in a row
     pub map: Vec<Vec<Option<Element>>>,
-    pub dimensions: Point<u32>,
+    pub dimensions: Point<i64>,
     pub default_element: Option<Element>,
 }
 
@@ -26,7 +26,7 @@ impl Layout {
     /// # Arguments
     ///
     /// * `dimensions` - The x-y dimensions of the map
-    pub fn new(dimensions: &Point<u32>, default_element: Option<Element>) -> Self {
+    pub fn new(dimensions: &Point<i64>, default_element: Option<Element>) -> Self {
         Self {
             map: vec![vec![default_element; dimensions.width as usize]; dimensions.height as usize],
             dimensions: *dimensions,
@@ -41,8 +41,8 @@ impl Layout {
 
         Self {
             dimensions: Point {
-                width: width as u32,
-                height: height as u32,
+                width: width as i64,
+                height: height as i64,
             },
             map,
             default_element,
@@ -95,9 +95,9 @@ impl Layout {
 
         let mut map = vec![vec![None; width as usize]; height as usize];
 
-        for index in 0..converted_items.len() as u32 {
-            let height = index / width;
-            let width = index % width;
+        for index in 0..converted_items.len() as i64 {
+            let height = index / width as i64;
+            let width = index % width as i64;
 
             map[height as usize][width as usize] = converted_items[index as usize];
         }
@@ -119,7 +119,7 @@ impl Layout {
         self
     }
 
-    pub fn get_row(&self, row_number: u32) -> MapResult<&Vec<Option<Element>>> {
+    pub fn get_row(&self, row_number: i64) -> MapResult<&Vec<Option<Element>>> {
         let row = self
             .map
             .get(row_number as usize)
@@ -129,7 +129,7 @@ impl Layout {
     }
 
     /// Returns the selected column_number. Has a differing type to get_row because we have to create an array of references and return it, whereas the get_row method returns a pointer to the row. A column doesn't exactly exist, it is just an element at the same row index for each row
-    pub fn get_column(&self, column_number: u32) -> MapResult<Vec<&Option<Element>>> {
+    pub fn get_column(&self, column_number: i64) -> MapResult<Vec<&Option<Element>>> {
         let mut items: Vec<&Option<Element>> = Vec::with_capacity(self.dimensions.height as usize);
 
         for height in 0..self.dimensions.height {
@@ -144,7 +144,7 @@ impl Layout {
         Ok(items)
     }
 
-    pub fn get_row_mut(&mut self, row_number: u32) -> MapResult<&mut Vec<Option<Element>>> {
+    pub fn get_row_mut(&mut self, row_number: i64) -> MapResult<&mut Vec<Option<Element>>> {
         let row = self
             .map
             .get_mut(row_number as usize)
@@ -153,7 +153,7 @@ impl Layout {
         Ok(row)
     }
 
-    // pub fn get_column_mut(&mut self, column_number: u32) -> MapResult<Vec<Option<&mut Element>>> {
+    // pub fn get_column_mut(&mut self, column_number: i64) -> MapResult<Vec<Option<&mut Element>>> {
     //     let mut items: Vec<Option<&mut Element>> =
     //         Vec::with_capacity(self.dimensions.height as usize);
 
@@ -166,7 +166,7 @@ impl Layout {
     //     Ok(items)
     // }
 
-    pub fn get_element_mut(&mut self, point: &Point<u32>) -> MapResult<&mut Option<Element>> {
+    pub fn get_element_mut(&mut self, point: &Point<i64>) -> MapResult<&mut Option<Element>> {
         let row = self.get_row_mut(point.height)?;
 
         let element = row
@@ -176,7 +176,7 @@ impl Layout {
         Ok(element)
     }
 
-    pub fn get_element(&self, point: &Point<u32>) -> MapResult<&Option<Element>> {
+    pub fn get_element(&self, point: &Point<i64>) -> MapResult<&Option<Element>> {
         let row = self.get_row(point.height)?;
 
         let element = row
@@ -188,8 +188,8 @@ impl Layout {
 
     pub fn draw_rect(
         &mut self,
-        start_position: &Point<u32>,
-        dimensions: &Point<u32>,
+        start_position: &Point<i64>,
+        dimensions: &Point<i64>,
         element: Element,
     ) -> Result<&mut Self, DisplayControllerError> {
         self.draw_line(
@@ -224,8 +224,8 @@ impl Layout {
     pub fn draw_line(
         &mut self,
         element: Element,
-        len: u32,
-        start_position: &Point<u32>,
+        len: i64,
+        start_position: &Point<i64>,
         direction: Direction,
     ) -> Result<&mut Self, DisplayControllerError> {
         match direction {
@@ -252,7 +252,7 @@ impl Layout {
     pub fn draw_item(
         &mut self,
         element: Element,
-        position: &Point<u32>,
+        position: &Point<i64>,
     ) -> Result<&mut Self, DisplayControllerError> {
         // Position is exclusive of the dimension borders
         if position.width >= self.dimensions.width || position.height >= self.dimensions.height {
@@ -273,10 +273,10 @@ mod tests {
 
     use super::Direction;
 
-    const WIDTH: u32 = 30;
-    const HEIGHT: u32 = 20;
+    const WIDTH: i64 = 30;
+    const HEIGHT: i64 = 20;
 
-    const DIMENSIONS: &Point<u32> = &Point::new(WIDTH, HEIGHT);
+    const DIMENSIONS: &Point<i64> = &Point::new(WIDTH, HEIGHT);
 
     const DEFAULT_ELEMENT: Element = Element::default();
